fix: looser schema; stricter type-narrowing

+++ static_report/src/components/shared/ColumnCard/ColumnTypeDetail/ColumnTypeDetailCategorical.tsx
@@ -40,7 +40,7 @@ export const ColumnTypeDetailCategorical: React.FC<Props> = ({
           value={formatModeMetrics(columnDatum)}
         />
       </Flex>
-      {type === 'integer' && (
+      {type === 'numeric' && (
         <>
           <Divider />
           <Flex mt={2}>
